requirements.txt
# JAIBird Stock Trading Platform Dependencies

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# Database
# sqlite3 is built into Python, no need to install
sqlalchemy==2.0.23

# Cloud Storage
dropbox==11.36.2

# Notifications
python-telegram-bot==20.7
# smtplib and email are built into Python

# Web Framework
flask==3.0.0
flask-wtf==1.2.1
wtforms==3.1.1

# Configuration Management
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI APIs
anthropic==0.7.8
openai==1.3.7

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# Utilities
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
# pathlib and logging are built into Python

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# AI & PDF Processing
openai==1.3.5
anthropic==0.7.8
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Telegram Bot
python-telegram-bot==20.6

# Performance & Monitoring
psutil==5.9.6
memory-profiler==0.61.0
